refactor: fix after review
