+++ src/components/rest-client/ResponsePanel.tsx
@@ -3,7 +3,7 @@
 import { Flex } from 'antd';
 
 import CodeSpace from '@/components/rest-client/CodeSpace';
-import { ApiResult, ResponseInfo } from '@/types/rest-client';
+import type { ApiResult, ResponseInfo } from '@/components/rest-client/types/rest-client';
 
 type Props = {
   result: ApiResult | undefined;
