+++ src/lib/antd/theme.ts
@@ -1,4 +1,4 @@
-import { ThemeConfig } from 'antd';
+import type { ThemeConfig } from 'antd';
 
 export const customTheme: ThemeConfig = {
   token: {
