+++ src/lib/store/selectors/counter.ts
@@ -1,6 +1,6 @@
 import { createSelector } from '@reduxjs/toolkit';
 
-import { RootState } from '../store';
+import type { RootState } from '../store';
 
 const getCounterState = (state: RootState) => state.counter;
